// types.ts
export type TContinentData = {
    i: string;    // Represents ISO code
    n: string;    // Represents name
    cn: any[];    // Represents countries (unused)
  };
  
export type TContinent = {
    iso: string;
    name: string;
};

export type TCountry = {
    iso: string;
    name: string;
};

export type TState = {
    iso: string;
    name: string;
};

export type TCity = {
    iso: string;
    name: string;
};


export interface ICountry {
	name: string;
	phonecode: string;
	isoCode: string;
	flag: string;
	currency: string;
	latitude: string;
	longitude: string;
	getAllCountries?(): ICountry[];
	getCountryByCode?(): ICountry;
}

export interface IState {
	name: string;
	isoCode: string;
	countryCode: string;
	latitude?: string | null;
	longitude?: string | null;
	getStatesOfCountry?(): IState[];
	getStateByCodeAndCountry?(): IState;
	getStateByCode?(): IState;
}
export interface ICity {
	name: string;
	countryCode: string;
	stateCode: string;
	latitude?: string | null;
	longitude?: string | null;
	getAllCities?(): ICity[];
	getCitiesOfState?(): ICity[];
	getCitiesOfCountry?(): ICity[];
}